requirements.txt
python-telegram-bot>=20.7
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=2.2.2
requests>=2.31.0
pyyaml>=6.0.1